refactor: change project structure of client app
