+++ client/src/features/header/Header.tsx
@@ -1,11 +1,11 @@
 import { useState } from 'react';
 import { Menu, X } from 'lucide-react';
 import { Link } from '@tanstack/react-router';
-import { SearchBar } from './SearchBar';
-import { AuthAction } from './AuthAction';
-import { MobileMenu } from './MobileMenu';
-import { ThemeToggle } from './ThemeToggle';
-import { Logo } from './Logo';
+import { SearchBar } from '@/features/header/components/SearchBar';
+import { AuthAction } from '@/features/header/components/AuthAction';
+import { MobileMenu } from '@/features/header/components/MobileMenu';
+import { ThemeToggle } from '@/features/header/components/ThemeToggle';
+import { Logo } from '@/features/header/components/Logo';
 
 const navLinkClassName =
   'text-[17px] font-serif font-normal text-foreground/85 hover:text-foreground transition-colors whitespace-nowrap';
