+++ client/src/features/header/components/SearchBar.tsx
