+++ client/src/features/opinions/index.ts
@@ -1,0 +1,4 @@
+import OpinionLabel from './OpinionLabel';
+import TableOfOpinions from './TableOfOpinions';
+
+export { OpinionLabel, TableOfOpinions };
