+++ client/src/features/subjects/index.ts
@@ -1,0 +1,3 @@
+import TableOfSubjects from './TableOfSubjects';
+
+export { TableOfSubjects };
