+++ client/src/hooks/__tests__/useAuth.test.tsx
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { act, renderHook } from '@testing-library/react';
-import { useAuth } from './useAuth';
+import { useAuth } from '../useAuth';
 import { AuthProvider } from '@/contexts/AuthProvider';
 
 function wrapper({
